uint64 -> int 2. Format package

+++ app/services/consumer/pkg/service/storage.go
@@ -7,15 +7,15 @@ import (
 // ConsumerStorage is the interface for the consumer storage.
 type ConsumerStorage interface {
 	InsertConsumer(consumer domain.Consumer) (*domain.Consumer, error)
-	DeleteConsumer(id uint64) error
+	DeleteConsumer(id int) error
 	UpdateConsumer(consumer domain.Consumer) (*domain.Consumer, error)
 	GetAllConsumer() ([]domain.Consumer, error)
-	GetConsumerByID(id uint64) (*domain.Consumer, error)
+	GetConsumerByID(id int) (*domain.Consumer, error)
 	GetConsumerDuplicateByParam(param domain.SearchParam) (*domain.Consumer, error)
 	CleanConsumerTable() error
 
 	InsertConsumerLocation(consumer domain.ConsumerLocation) (*domain.ConsumerLocation, error)
-	DeleteConsumerLocation(consumerID uint64) error
+	DeleteConsumerLocation(consumerID int) error
 	UpdateConsumerLocation(consumer domain.ConsumerLocation) (*domain.ConsumerLocation, error)
-	GetConsumerLocation(id uint64) (*domain.ConsumerLocation, error)
+	GetConsumerLocation(id int) (*domain.ConsumerLocation, error)
 }
